Checking authentication & authorization on each API request, so that only authenticated and authorized user can see the results

+++ ToDoApp/routers/todos.py
@@ -7,6 +7,7 @@ from starlette import status
 
 from models import Todos
 from database import SessionLocal
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -20,23 +21,28 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str = Field(min_length=3, max_length=100)
     priority: int = Field(gt=0, lt=6)
-    complete: bool
+    complete: bool = Field(default=False)
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if not user:
+        raise HTTPException(status_code=401, detail='Authentication Failed.')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=401, detail='Authentication Failed.')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model:
         return todo_model
     else:
@@ -44,16 +50,21 @@ async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_req: TodoRequest):
-    todo_model = Todos(**todo_req.dict())
+async def create_todo(user: user_dependency, db: db_dependency, todo_req: TodoRequest):
+    if not user:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = Todos(**todo_req.dict(), owner_id=user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_req: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_req: TodoRequest, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=401, detail='Authentication Failed.')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
     if not todo_model:
         raise HTTPException(status_code=404, detail=f"todo_id {todo_id}, not found.")
@@ -68,12 +79,15 @@ async def update_todo(db: db_dependency, todo_req: TodoRequest, todo_id: int = P
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=401, detail='Authentication Failed.')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
     if not todo_model:
         raise HTTPException(status_code=404, detail=f"todo_id {todo_id}, not found.")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id, Todos.owner_id == user.get('id')).delete()
 
     db.commit()
